rf: split out screens into separate components, css
