+++ src/App.tsx
@@ -1,4 +1,4 @@
-import Game from 'components/Game'
+import Game from 'components/Game/Game'
 
 const App = () => {
   return <Game />
